Make localization easier and more flexible
Current localization framework is not flexible and has a few drawbacks.
Mainly
* It places translations in the source code, which makes localization almost
  impossible for a casual translator.
* Localized strings make the source code lengthy and hard to maintain.
* Placing strings in different encodings in a single file is clumsy
  and can be easily broken while editing.

The following change improves localizaton both for developers and translators.
The code itself is not cluttered with language translations and thus easier to
maintain for developers. Adding a new language (or editing a string) is an easy
task for a translator - there is no need to mess with code.

Translated strings are extracted to separate text files, separately for different
languages and encodings. A simple framework for setting strings is introduced
to generate C sources from the language files by make.

Note:
It might be better to use gettext() or another localization tool, instead.
But this is an improvement, anyway.